requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
appdirs==1.4.4
    # via
    #   -r requirements/base.txt
    #   fs
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
boto3==1.7.0
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   fs-s3fs
botocore==1.10.84
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   boto3
    #   s3transfer
click==8.1.7
    # via code-annotations
code-annotations==1.5.0
    # via -r requirements/test.in
coverage[toml]==7.3.0
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-appconf
    #   django-model-utils
    #   django-statici18n
    #   djangorestframework
    #   edx-i18n-tools
    #   edx-submissions
    #   jsonfield
    #   openedx-django-pyfs
django-appconf==1.0.5
    # via
    #   -r requirements/base.txt
    #   django-statici18n
django-model-utils==4.3.1
    # via
    #   -r requirements/base.txt
    #   edx-submissions
django-statici18n==2.4.0
    # via -r requirements/base.txt
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   edx-submissions
docutils==0.20.1
    # via
    #   -r requirements/base.txt
    #   botocore
edx-i18n-tools==1.1.0
    # via -r requirements/base.txt
edx-submissions==3.6.0
    # via -r requirements/base.txt
exceptiongroup==1.1.3
    # via pytest
fs==2.4.16
    # via
    #   -r requirements/base.txt
    #   fs-s3fs
    #   openedx-django-pyfs
    #   xblock
fs-s3fs==1.1.0
    # via
    #   -r requirements/base.txt
    #   openedx-django-pyfs
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
jmespath==0.10.0
    # via
    #   -r requirements/base.txt
    #   boto3
    #   botocore
jsonfield==3.1.0
    # via
    #   -r requirements/base.txt
    #   edx-submissions
lazy==1.5
    # via
    #   -r requirements/base.txt
    #   xblock
lxml==4.9.3
    # via
    #   -r requirements/base.txt
    #   xblock
mako==1.2.4
    # via
    #   -r requirements/base.txt
    #   xblock-utils
markupsafe==2.1.3
    # via
    #   -r requirements/base.txt
    #   jinja2
    #   mako
    #   xblock
openedx-django-pyfs==3.4.0
    # via
    #   -r requirements/base.txt
    #   xblock
packaging==23.1
    # via pytest
path==16.7.1
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pbr==5.11.1
    # via stevedore
pluggy==1.2.0
    # via pytest
polib==1.2.0
    # via
    #   -r requirements/base.txt
    #   edx-i18n-tools
pytest==7.4.0
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   botocore
    #   xblock
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
    #   edx-submissions
    #   xblock
pyyaml==6.0.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-i18n-tools
    #   xblock
s3transfer==0.1.13
    # via
    #   -r requirements/base.txt
    #   boto3
simplejson==3.19.1
    # via
    #   -r requirements/base.txt
    #   xblock-utils
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   fs
    #   fs-s3fs
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.7.1
    # via
    #   -r requirements/base.txt
    #   asgiref
web-fragments==2.1.0
    # via
    #   -r requirements/base.txt
    #   xblock
    #   xblock-utils
webob==1.8.7
    # via
    #   -r requirements/base.txt
    #   xblock
xblock[django]==1.7.0
    # via
    #   -r requirements/base.txt
    #   xblock-utils
xblock-utils==3.4.1
    # via -r requirements/base.txt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
